- New dial-knob behavior now follows mouse pointer angular position,   almost similar to old QDial, but this time avoiding that nasty and   rather abrupt change on first mouse click. (FIX)
git-svn-id: http://svn.code.sf.net/p/qsynth/code_attic/trunk@184 6e21f190-4f5b-40a6-a6af-ed3d11dd9ba7

+++ src/qsynthKnob.h
@@ -71,11 +71,11 @@ public slots:
 
 protected:
 
-	void drawTick(QPainter& paint, float angle, int size, bool internal);
+	void drawTick(QPainter& paint, double angle, int size, bool internal);
 	virtual void repaintScreen(const QRect *pRect = 0);
 
 	// Mouse angle determination.
-	float mouseAngle(const QPoint& pos);
+	double mouseAngle(const QPoint& pos);
 
 	// Alternate mouse behavior event handlers.
 	virtual void mousePressEvent(QMouseEvent *pMouseEvent);
